requirements.txt
cloudscraper
beautifulsoup4
colorama
PyJWT
